+++ include/phtree/common/b_plus_tree_hash_map.h
@@ -83,12 +83,10 @@ class b_plus_tree_hash_set {
 
     class bpt_node_leaf;
     class bpt_iterator;
-    using LeafEntryT = std::pair<hash_t, T>;
     using IterT = bpt_iterator;
     using NLeafT = bpt_node_leaf;
-    using NInnerT = bpt_node_inner<hash_t, NLeafT, IterT>;
+    using NInnerT = bpt_node_inner<hash_t, NLeafT>;
     using NodeT = bpt_node_base<hash_t, NInnerT, bpt_node_leaf>;
-    using LeafIteratorT = decltype(std::vector<LeafEntryT>().begin());
     using TreeT = b_plus_tree_hash_set<T, HashT, PredT>;
 
   public:
@@ -241,7 +239,7 @@ class b_plus_tree_hash_set {
     }
 
   private:
-    using bpt_leaf_super = bpt_node_data<hash_t, NInnerT, NLeafT, NLeafT, LeafEntryT, IterT>;
+    using bpt_leaf_super = bpt_node_data<hash_t, T, NInnerT, NLeafT, true>;
     class bpt_node_leaf : public bpt_leaf_super {
       public:
         explicit bpt_node_leaf(NInnerT* parent, NLeafT* prev, NLeafT* next) noexcept
@@ -251,7 +249,7 @@ class b_plus_tree_hash_set {
 
         [[nodiscard]] IterT find(hash_t hash, const T& value) noexcept {
             PredT equals{};
-            IterT iter_full = this->lower_bound_as_iter(hash);
+            IterT iter_full = this->template lower_bound_as_iter<IterT>(hash);
             while (!iter_full.is_end() && iter_full.hash() == hash) {
                 if (equals(*iter_full, value)) {
                     return iter_full;
@@ -263,7 +261,7 @@ class b_plus_tree_hash_set {
 
         [[nodiscard]] auto lower_bound_value(hash_t hash, const T& value) noexcept {
             PredT equals{};
-            IterT iter_full = this->lower_bound_as_iter(hash);
+            IterT iter_full = this->template lower_bound_as_iter<IterT>(hash);
             while (!iter_full.is_end() && iter_full.hash() == hash) {
                 if (equals(*iter_full, value)) {
                     break;
@@ -288,7 +286,7 @@ class b_plus_tree_hash_set {
             }
             ++entry_count;
 
-            auto full_it = this->check_split_and_adjust_iterator(it, hash, root);
+            auto full_it = this->template check_split_and_adjust_iterator<IterT>(it, hash, root);
             auto it_result = full_it.node_->data_.emplace(full_it.iter_, hash, std::move(t));
             return std::make_pair(IterT(full_it.node_, it_result), true);
         }
@@ -312,8 +310,8 @@ class b_plus_tree_hash_set {
         }
     };
 
-    class bpt_iterator : public bpt_iterator_base<LeafIteratorT, NLeafT, NodeT, TreeT> {
-        using SuperT = bpt_iterator_base<LeafIteratorT, NLeafT, NodeT, TreeT>;
+    class bpt_iterator : public bpt_iterator_base<NLeafT, NodeT, TreeT> {
+        using SuperT = bpt_iterator_base<NLeafT, NodeT, TreeT>;
 
       public:
         using iterator_category = std::forward_iterator_tag;
@@ -323,7 +321,8 @@ class b_plus_tree_hash_set {
         using reference = T&;
 
         // Arbitrary position iterator
-        explicit bpt_iterator(NLeafT* node, LeafIteratorT it) noexcept : SuperT(node, it) {}
+        explicit bpt_iterator(NLeafT* node, typename SuperT::LeafIteratorT it) noexcept
+        : SuperT(node, it) {}
 
         // begin() iterator
         explicit bpt_iterator(NodeT* node) noexcept : SuperT(node) {}