+++ include/phtree/common/b_plus_tree_map.h
@@ -88,12 +88,10 @@ class b_plus_tree_map {
 
     class bpt_node_leaf;
     class bpt_iterator;
-    using LeafEntryT = std::pair<KeyT, ValueT>;
     using IterT = bpt_iterator;
     using NLeafT = bpt_node_leaf;
-    using NInnerT = bpt_node_inner<KeyT, NLeafT, IterT, INNER_CFG>;
+    using NInnerT = bpt_node_inner<KeyT, NLeafT, INNER_CFG>;
     using NodeT = bpt_node_base<KeyT, NInnerT, bpt_node_leaf>;
-    using LeafIteratorT = decltype(std::vector<LeafEntryT>().begin());
     using TreeT = b_plus_tree_map<KeyT, ValueT, COUNT_MAX>;
 
   public:
@@ -142,7 +140,7 @@ class b_plus_tree_map {
 
     [[nodiscard]] auto lower_bound(KeyT key) noexcept {
         auto leaf = lower_bound_leaf(key, root_);
-        return leaf != nullptr ? leaf->lower_bound_as_iter(key) : IterT{};
+        return leaf != nullptr ? leaf->template lower_bound_as_iter<IterT>(key) : IterT{};
     }
 
     [[nodiscard]] auto lower_bound(KeyT key) const noexcept {
@@ -166,6 +164,10 @@ class b_plus_tree_map {
     }
 
     [[nodiscard]] auto end() const noexcept {
+        return IterT();
+    }
+
+    [[nodiscard]] auto cend() const noexcept {
         return IterT();
     }
 
@@ -231,8 +233,7 @@ class b_plus_tree_map {
     }
 
   private:
-    using bpt_leaf_super =
-        bpt_node_data<KeyT, NInnerT, NLeafT, NLeafT, LeafEntryT, IterT, LEAF_CFG>;
+    using bpt_leaf_super = bpt_node_data<KeyT, ValueT, NInnerT, NLeafT, true, LEAF_CFG>;
     class bpt_node_leaf : public bpt_leaf_super {
       public:
         explicit bpt_node_leaf(NInnerT* parent, NLeafT* prev, NLeafT* next) noexcept
@@ -256,7 +257,7 @@ class b_plus_tree_map {
             }
             ++entry_count;
 
-            auto full_it = this->check_split_and_adjust_iterator(it, key, root);
+            auto full_it = this->template check_split_and_adjust_iterator<IterT>(it, key, root);
             auto it_result = full_it.node_->data_.emplace(
                 full_it.iter_,
                 std::piecewise_construct,
@@ -289,8 +290,8 @@ class b_plus_tree_map {
         }
     };
 
-    class bpt_iterator : public bpt_iterator_base<LeafIteratorT, NLeafT, NodeT, TreeT> {
-        using SuperT = bpt_iterator_base<LeafIteratorT, NLeafT, NodeT, TreeT>;
+    class bpt_iterator : public bpt_iterator_base<NLeafT, NodeT, TreeT> {
+        using SuperT = bpt_iterator_base<NLeafT, NodeT, TreeT>;
 
       public:
         using iterator_category = std::forward_iterator_tag;
@@ -300,7 +301,8 @@ class b_plus_tree_map {
         using reference = ValueT&;
 
         // Arbitrary position iterator
-        explicit bpt_iterator(NLeafT* node, LeafIteratorT it) noexcept : SuperT(node, it) {}
+        explicit bpt_iterator(NLeafT* node, typename SuperT::LeafIteratorT it) noexcept
+        : SuperT(node, it) {}
 
         // begin() iterator
         explicit bpt_iterator(NodeT* node) noexcept : SuperT(node) {}
@@ -309,11 +311,11 @@ class b_plus_tree_map {
         bpt_iterator() noexcept : SuperT() {}
 
         auto& operator*() const noexcept {
-            return const_cast<LeafEntryT&>(*this->iter());
+            return *this->iter();
         }
 
         auto* operator->() const noexcept {
-            return const_cast<LeafEntryT*>(&*this->iter());
+            return &*this->iter();
         }
     };